+++ backend/db/models/event.js
@@ -65,7 +65,7 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false
     },
     price: {
-      type: DataTypes.DECIMAL(5, 1)
+      type: DataTypes.FLOAT(5, 1)
     },
     previewImage: {
       type: DataTypes.STRING
